+++ module_src/verify_functions_extends/general-cinnamonBun/verify_functions.sh
@@ -24,46 +24,46 @@ verify_android_api_has_pass() {
 
 verify_special_rule_pass() {
   local mi_os_version_code=$(grep_prop ro.mi.os.version.code /mi_ext/etc/build.prop)
-  local sothx_project_treble_support_autoui_services_fix=$(getprop ro.config.sothx_project_treble_support_autoui_services_fix)
-  local sothx_project_treble_autoui_services_fix_version=$(getprop ro.config.sothx_project_treble_autoui_services_fix_version)
+  local sothx_project_treble_support_hyper_magic_window_cloud_config_fix=$(getprop ro.config.sothx_project_treble_support_hyper_magic_window_cloud_config_fix)
+  local sothx_project_treble_hyper_magic_window_cloud_config_fix_version=$(getprop ro.config.sothx_project_treble_hyper_magic_window_cloud_config_fix_version)
   local sothx_disabled_os2_install_module_tips=$(getprop ro.sothx.disabled_os2_install_module_tips)
-  local is_need_install_autoui_cloudfix_apk=1
-  HAS_BEEN_INSTALLED_AutoUI_CloudFix_APK=$(pm list packages | grep io.github.sothx.autouicloudfix)
-  if [[ $HAS_BEEN_INSTALLED_AutoUI_CloudFix_APK == *"package:io.github.sothx.autouicloudfix"* ]]; then
-    is_need_install_autoui_cloudfix_apk=0
+  local is_need_install_fix_hyper_magic_window_cloud_config_apk=1
+  HAS_BEEN_INSTALLED_FIX_HYPER_MAGIC_WINDOW_CLOUD_CONFIG_APK=$(pm list packages | grep io.github.sothx.FixHyperMagicWindowCloudConfig)
+  if [[ $HAS_BEEN_INSTALLED_FIX_HYPER_MAGIC_WINDOW_CLOUD_CONFIG_APK == *"package:io.github.sothx.FixHyperMagicWindowCloudConfig"* ]]; then
+    is_need_install_fix_hyper_magic_window_cloud_config_apk=0
   fi
-  if [[ $sothx_project_treble_support_autoui_services_fix == 'true' ]]; then
-    if [[ $sothx_project_treble_autoui_services_fix_version -ge 2 ]]; then
-      is_need_install_autoui_cloudfix_apk=0
+  if [[ $sothx_project_treble_support_hyper_magic_window_cloud_config_fix == 'true' ]]; then
+    if [[ $sothx_project_treble_hyper_magic_window_cloud_config_fix_version -ge 1 ]]; then
+      is_need_install_fix_hyper_magic_window_cloud_config_apk=0
     fi
   fi
 
-  if [[ $is_need_install_autoui_cloudfix_apk == 1 ]]; then
+  if [[ $is_need_install_fix_hyper_magic_window_cloud_config_apk == 1 ]]; then
     ui_print "*********************************************"
-    ui_print "- 是否安装AutoUI Cloud Fix？"
-    ui_print "- [重要提醒]: Hyper OS 4 需要安装该LSPosed模块才能修复应用布局优化不生效的问题"
+    ui_print "- 是否安装Fix Hyper Magic Window Cloud Config？"
+    ui_print "- [重要提醒]: Hyper OS 4 需要安装该LSPosed模块才能修复应用横屏布局和应用布局优化不生效的问题"
     ui_print "- [重要提醒]: 安装后请前往LSPosed启用模块并重启平板"
     ui_print "  音量+ ：是"
     ui_print "  音量- ：否"
     ui_print "*********************************************"
     key_check
     if [[ "$keycheck" == "KEY_VOLUMEUP" ]]; then
-      ui_print "- 正在为你安装AutoUI Cloud Fix，请稍等~"
-      unzip -jo "$ZIPFILE" 'common/apks/AutoUICloudFix.apk' -d /data/local/tmp/ &>/dev/null
-      pm install -r /data/local/tmp/AutoUICloudFix.apk &>/dev/null
-      rm -rf /data/local/tmp/AutoUICloudFix.apk
-      HAS_BEEN_INSTALLED_AutoUI_CloudFix_APK=$(pm list packages | grep io.github.sothx.autouicloudfix)
-      if [[ $HAS_BEEN_INSTALLED_AutoUI_CloudFix_APK == *"package:io.github.sothx.autouicloudfix"* ]]; then
-        ui_print "- 好诶，AutoUI Cloud Fix安装完成！"
+      ui_print "- 正在为你安装Fix Hyper Magic Window Cloud Config，请稍等~"
+      unzip -jo "$ZIPFILE" 'common/apks/FixHyperMagicWindowCloudConfig.apk' -d /data/local/tmp/ &>/dev/null
+      pm install -r /data/local/tmp/FixHyperMagicWindowCloudConfig.apk &>/dev/null
+      rm -rf /data/local/tmp/FixHyperMagicWindowCloudConfig.apk
+      HAS_BEEN_INSTALLED_FIX_HYPER_MAGIC_WINDOW_CLOUD_CONFIG_APK=$(pm list packages | grep io.github.sothx.FixHyperMagicWindowCloudConfig)
+      if [[ $HAS_BEEN_INSTALLED_FIX_HYPER_MAGIC_WINDOW_CLOUD_CONFIG_APK == *"package:io.github.sothx.FixHyperMagicWindowCloudConfig"* ]]; then
+        ui_print "- 好诶Fix Hyper Magic Window Cloud Config安装完成！"
         ui_print "- 好诶，安装后请前往LSPosed启用模块并重启平板！"
       else
-        abort "- AutoUI Cloud Fix安装失败，请尝试重新安装！"
-        abort "- 也可前往模块网盘下载单独的 AutoUI Cloud Fix apk 进行手动安装！"
+        abort "- Fix Hyper Magic Window Cloud Config安装失败，请尝试重新安装！"
+        abort "- 也可前往模块网盘下载单独的 Fix Hyper Magic Window Cloud Config apk 进行手动安装！"
       fi
     else
       ui_print "*********************************************"
-      ui_print "- 你选择不安装AutoUI Cloud Fix"
-      ui_print "- Hyper OS 4 需要安装该LSPosed模块才能修复应用布局优化不生效的问题"
+      ui_print "- 你选择不安装Fix Hyper Magic Window Cloud Config"
+      ui_print "- Hyper OS 4 需要安装该LSPosed模块才能修复应用横屏布局和应用布局优化不生效的问题"
       ui_print "*********************************************"
     fi
   fi